pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ciguard"
version = "0.1.0"
description = "Static security auditor for CI/CD pipelines (GitLab CI today; GitHub Actions + Jenkins in development)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [{ name = "Jo-Jo98" }]
keywords = ["ci", "cd", "security", "sast", "gitlab", "audit", "compliance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "reportlab>=4.0",
    "anthropic>=0.34",
    "openai>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/Jo-Jo98/ciguard"
Repository = "https://github.com/Jo-Jo98/ciguard"
Issues = "https://github.com/Jo-Jo98/ciguard/issues"

[project.scripts]
ciguard = "ciguard.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ciguard = ["reporter/templates/*.html", "web/templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py39"
